Partitioned codebase into api.py (rough API and utilities) and sdowebapp.py (HTTP/HTML web app). Updated supporting files and tests accordingly. The API needs rationalizing, but is a start at least on removing some spaghetti. Eventually more can be exposed via templates/*tpl.
